Rename get-api-key-credit-v2 core module
