+++ notes/agents/2026-07-06-get-api-key-credit-v2-core-rename.md
@@ -1,0 +1,4 @@
+# Agent Retrospective: get-api-key-credit-v2 core rename
+
+- **Challenge:** Renaming the `get-api-key-credit-v2` core module required updating both the cloud re-export and the copy-to-infra script so deployments receive the implementation under the new filename.
+- **Resolution:** Renamed the source files, updated all imports and tests to point at `get-api-key-credit-v2-core.js`, and adjusted the copy script to mirror the renamed module into the Cloud Function bundle.
