--- src/cloud/get-api-key-credit-v2/core.js
@@ -1,1 +1,0 @@
-export * from '../../core/cloud/get-api-key-credit-v2/core.js';
